+++ studio/pages/project/[ref]/database/backups/scheduled.tsx
@@ -1,16 +1,20 @@
 import { useRouter } from 'next/router'
 import { observer } from 'mobx-react-lite'
 import { Tabs } from '@supabase/ui'
+import { PermissionAction } from '@supabase/shared-types/out/constants'
 
-import { useStore } from 'hooks'
+import { NextPageWithLayout } from 'types'
+import { checkPermissions, useStore } from 'hooks'
 import { DatabaseLayout } from 'components/layouts'
 import { BackupsList } from 'components/interfaces/Database'
-import { NextPageWithLayout } from 'types'
+import NoPermission from 'components/ui/NoPermission'
 
 const DatabaseScheduledBackups: NextPageWithLayout = () => {
   const { ui } = useStore()
   const router = useRouter()
   const ref = ui.selectedProject?.ref
+
+  const canReadScheduledBackups = checkPermissions(PermissionAction.READ, 'back_ups')
 
   return (
     <div className="mx-auto max-w-4xl space-y-6 px-5 pt-12 pb-20">
@@ -32,7 +36,11 @@ const DatabaseScheduledBackups: NextPageWithLayout = () => {
           Projects are backed up daily around midnight of your project's region and can be restored
           at any time.
         </p>
-        <BackupsList />
+        {canReadScheduledBackups ? (
+          <BackupsList />
+        ) : (
+          <NoPermission resourceText="view scheduled backups" />
+        )}
       </div>
     </div>
   )
